Moved cleaner to its own namespace.
