+++ classes/local/cleaner/cleaner.php
@@ -23,11 +23,12 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace local_cleanurls;
+namespace local_cleanurls\local\cleaner;
 
 use cache;
 use cache_application;
 use cm_info;
+use local_cleanurls\clean_moodle_url;
 use moodle_url;
 use stdClass;
